requirements.txt
pyTelegramBotAPI
flask
requests
python-dotenv
wit